clarify wording about current working directory
"in the current working directory" implies that the contents of the
rendered template will be placed directly into the current working
directory, as though the cwd were the root of the project.

Changing "in" to "under" I think more accurately communicates that the
root of the new project repo will be created within the cwd.