change where unittest.xml is generated
PLC-3440

from pytest's standpoint, the current working directory is something
under .pybuild/ which is a staging area for building the .deb. This
results in the .deb deploying
/usr/lib/python3/dist-packages/unittest.xml to the filesystem. Because
it will be the same path for each repository, apt will complain about a
file already owned by another package.

CURDIR is the current working directory where make was first invoked
(before it recurses). This will be the top level of the repo. For a
developer, this is the root of the working copy of their git clone. For
jenkins this is some clone of the repo in /tmp/... in the cowbuilder
chroot.