implement original Makefile
Using `-include` instead of `include` is important for jenkins, which
does not necessarily have gwn-dev-tools installed to provide the common
make file inside the build environment. `debian/rules` will *always*
result in the Makefile being processed, even though it's strictly for
developer convenience. The leading hyphen makes non-strict, and so the
missing common Makefile won't fail the build.

Not sure about the lack of the body to the `default` target, but it was
in Will's original implementation, and I've seen make issuing warnings
as I've played around with Makefile syntax.